requirements.txt
Django==3.2.8
django-cors-headers==3.7.0
bcrypt==3.2.0
PyJWT==2.1.0
requests==2.25.1